+++ src/parse_utils.cpp
@@ -12,7 +12,7 @@
  * @param output_len Length to be stored in output buffer
  * @return  size of data stored in output_buf
  */
-ssize_t  extract_from_buffer (unsigned char*& buffer, int &buf_len, void *output_buf, int output_len) {
+uint32_t extract_from_buffer (unsigned char*& buffer, int &buf_len, void *output_buf, int output_len) {
     if (output_len > buf_len)
         return (output_len - buf_len);
     memcpy(output_buf, buffer, output_len);
